dataset() class upgrade -- .plotContorno barely available
add plotContorno for ploting 2D contour of dataset. this version plots a heatmap of all probabilities combined to a single image. maybe define a self.pdfSurface for persistent storage of (class & probability) informaton.
TODO:  surface classifier  : this implies some classification rule. In this case is required that a classification boundary